requirements.txt
fastapi==0.115.12
joblib==1.4.2
openai==1.71.0
pandas==2.2.3
pydantic==2.11.3
python-dotenv==1.1.0
scikit_learn==1.6.1
spacy==3.8.5
tensorflow==2.15.0
tensorflow_macos==2.15.0
uvicorn==0.34.0
